requirements.txt
python-dotenv==1.0.0
keyring==24.3.0 